+++ docs/source/changelog.rst
@@ -18,6 +18,7 @@ Change Log
 - socket SO_REUSEADDR option now not set anymore by default; added new config item SOCK_REUSE to be able to set it to True if you want.
 - threaded server should deal with EINTR and other errors better (retry call)
 - better closedown of threadpool server
+- fix for potential autoproxy failure when unregistering pyro objects
 
 
 **Pyro 4.8**